+++ s4e-web/src/app/views/map-view/view-manager/view-manager.component.ts
@@ -98,7 +98,7 @@ export class ViewManagerComponent implements OnInit, OnDestroy{
   }
 
 
-  isFavoriteProduct = (ID: number, isFavorite: boolean): void => this.productService.toggleFavorite(ID, isFavorite);
+  isFavouriteProduct = (ID: number, isFavourite: boolean): void => this.productService.toggleFavourite(ID, isFavourite);
 
   searchForPlaces(place: string) {
     this.searchResultsService.get(place);